requirements.txt
click
ffmpeg-python
ultralytics